v0.13.9 (#510)
New Features:
- Add default representation for objects without a `tostr` method.
- Add variadic list assigns to declarations (to match already supported syntax for functions).
- Add wildcard matching via `str.like`.
- Add `debug.echobacktrace()`, to print out a stack trace. The exact format is subject to change.
- Add `debug.getglobal` and `debug.setglobal`, for inspecting/manipulating global variables.
- Add `debug.getlocal` and `debug.setlocal` for inspecting/manipulating values from a frame.
- Allow `math.seed` with no argument to set us into a new random state.

Bug Fixes:
- Fix formatting for stringifying functions.
- Fix removing non-existent element from `table`.
- Fix removing non-existent element from `set`.
- `str.tobyte` always returns a non-negative `int` now.
- Fix error case of `list.removeindex`.
- Fix invalid read in `list.remove` when size equals capacity.
- Fix behaviour of `str.has`, `str.search`, and `str.searchall` with empty `str`.
- Fix behaviour of `list.has`, `list.search`, and `list.searchall` with empty `list`.
- Fix error message for `list.shuffle`.
- Fix handling of invalid inputs for `list.__next` and `str.__next`.
- Fix `echo` with `str`s containing a `\0` character.

Internal Changes:
- Improve hash function used for `str`.
- Switch to custom PRNG.

+++ src/compiler/lexer.c
@@ -221,7 +221,7 @@ static long lex_eatnumber_fill(struct Lexer *const lex, int (*isvaliddigit)(int)
 }
 
 static bool lex_eatint(struct Lexer *const lex, char separator, int (*isvaliddigit)(int)) {
-	int curr_pos = lxtell(lex->file);
+	long curr_pos = lxtell(lex->file);
 	int curr_char = lex->c;
 	if (lex->c == '0') {
 		lex_getchar(lex);
